refactor(backend): connection_utils を 4 ファイルに責務分割 (#454)
