+++ backend/database/async_connection_executor.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "../utils/transparent_hash.h"
-#include "connection_utils.h"
+#include "connection_types.h"
 #include "driver_interface.h"
 
 #include <atomic>
