+++ backend/providers/schema_provider.cpp
@@ -1,6 +1,6 @@
 #include "schema_provider.h"
 
-#include "../database/connection_utils.h"
+#include "../database/connection_params_parser.h"
 #include "../database/driver_interface.h"
 #include "../interfaces/ddl_queryable.h"
 #include "../interfaces/providers/connection_provider.h"
